started BookingController and modified BookingDAO

+++ src/main/java/DomainModel/Booking.java
@@ -7,12 +7,12 @@ import java.util.Date;
 public class Booking {
     private int ID;
     private LocalDate date;
-    private int period; //duration of the booking
+    private float period; //duration of the booking
     private LocalTime time;
     private User user;
     private Field field;
 
-    public Booking(int ID, LocalDate date, int period,LocalTime time, User user,Field field){
+    public Booking(int ID, LocalDate date, float period,LocalTime time, User user,Field field){
         this.ID=ID;
         this.date=date;
         this.period=period;
@@ -22,7 +22,7 @@ public class Booking {
     }
     public int getID(){return ID;}
     public LocalDate getDate (){return date;}
-    public int getPeriod(){return period;}
+    public float getPeriod(){return period;}
     public LocalTime getTime(){return time;}
     public User getUser() {
         return user;
